refactor: Entity 패키지 분리
- Car, Cars 클래스에 대해 Entity 패키지로 분리
