+++ src/test/java/racingcar/model/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
