+++ src/test/java/racingcar/model/CarsTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import java.util.List;
 import org.junit.jupiter.api.Assertions;
